moving fxa away from routers
