--- src/mlpa/core/routers/fxa/__init__.py
@@ -1,5 +1,0 @@
-from mlpa.core.routers.fxa.fxa import fxa_auth
-
-__all__ = [
-    "fxa_auth",
-]
