+++ src/tests/integration/conftest.py
@@ -45,7 +45,7 @@ def mocked_client_integration(mocker, use_real_get_or_create_user):
     mocker.patch("mlpa.core.routers.health.health.litellm_pg", mock_litellm_pg)
     mocker.patch("mlpa.core.utils.litellm_pg", mock_litellm_pg)
 
-    mocker.patch("mlpa.core.routers.fxa.fxa.client", mock_fxa_client)
+    mocker.patch("mlpa.core.auth.fxa.client", mock_fxa_client)
 
     mocker.patch(
         "mlpa.core.routers.mock.mock.fxa_client", mock_fxa_client_for_mock_router
